refactor(ctxpkg): split verifier for LOC gate
